Added cast to avoid warning from GCC 4. The length will always be shorter than will fit in 32 bits, so treating as an int is fine.

+++ src/Exporters.cxx
@@ -943,7 +943,7 @@ void SciTEBase::SaveToPDF(const char *saveName) {
 			// concatenate stream within the text object
 			sprintf(textObj, "<</Length %d>>\nstream\n%s"
 					 "ET\nendstream\n",
-					 pageData.length() - 1 + 3,
+					 static_cast<int>(pageData.length() - 1 + 3),
 					 pageData.c_str());
 			oT->add(textObj);
 			delete []textObj;
